requirements.txt
pydantic
pendulum
kafka-python
jinja2
fastapi
uvicorn
pyyaml
python-dotenv